Only check the bound bits of encodings
Encoding sets contain a mix of encodings, for which they have a differing set of bound bits. We sort them by decreasing number of bound bits. We then check them, one by one, against `ir`. Given they have a differing set of bound bits, we only want to be checking that bits that each encoding have bound; it is the ordering that ensures we don't unduly select an encoding that shadows another.